+++ src/backtesting/display.py
